requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
docopt==0.6.2
Flask==3.1.1
fonttools==4.58.4
grip==4.6.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
path-and-address==2.0.1
pathlib==1.0.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.2
pyopencl==2025.2.4
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
pytools==2025.1.6
PyYAML==6.0.2
requests==2.32.4
siphash24==1.7
six==1.17.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
